split ui.js apart
